done paination

+++ core/models.py
@@ -4,7 +4,7 @@ from django.db import models
 
 class Product(models.Model):
     name = models.CharField(max_length=100)
-    price = models.IntegerField()
+    price = models.DecimalField(decimal_places=2,max_digits=9999)
     quantity = models.IntegerField()
     image = models.ImageField(upload_to='pictures/')
     created_at = models.DateTimeField(auto_now_add=True)
